fix: Mark posts as read when opening expandos in Compact view
Do NOT check the visited state of each URL in the feed. Browsers de-prioritise and restrict extension background scripts from handling complex tasks, doing so will slow the function to a halt.

Instead, if the Mark read button is pressed, add the URL to the browser history again. Let users see if the URL is marked visited by themselves.

Also: internally merge "Add a button to each post to mark it as read" with markReadOnOpenExpandos, scan the feed with only one MutationObserver to reduce performance load on browser.

TODO: check if this closes #201.
Signed-off-by: Bùi Minh Đức <93396463+bmndc@users.noreply.github.com>